Use our player to be retained by the view controller
This avoids losing the player when only the picture in picture layer is alive
