--- Sources/Player/Extensions/AVPictureInPictureController.swift
@@ -1,13 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-import AVKit
-
-extension AVPictureInPictureController {
-    var parentPlayer: Player? {
-        contentSource?.playerLayer?.player?.parent
-    }
-}
