+++ Sources/Player/Player.swift
@@ -213,7 +213,6 @@ public final class Player: ObservableObject, Equatable {
         nowPlayingSession = MPNowPlayingSession(players: [queuePlayer])
 #endif
         self.configuration = configuration
-        queuePlayer.parent = self
 
         configurePlayer()
         configurePublishedPropertyPublishers()
